fix: bring trend column UI under eslint budgets
Split ColumnRow format fields into helpers and move progress/trend
normalize specs out of the oversized panelTypes.spec file.
